ai-book-publisher/requirements.txt
streamlit
bs4
requests
nltk
pyttsx3
chromadb
playwright
sentence-transformers
google-generativeai
python-dotenv